Added a middleware
Adding a middleware layer keeps the vulnerability processing seperate from the actual lanaguage server which should only be dealing with simple LSP requests
