+++ src/lsp/mod.rs
@@ -3,4 +3,5 @@ pub(crate) mod diagnostics;
 
 pub(crate) mod hover;
 pub(crate) mod language_server;
+pub(crate) mod middleware;
 pub(crate) mod progress;
